+++ libraries/bio/src/monomer-works/monomer-utils.ts
@@ -7,21 +7,20 @@ import {
   MONOMER_ENCODE_MAX, MONOMER_ENCODE_MIN, SDF_MONOMER_NAME
 } from '../utils/const';
 import {IMonomerLib} from '../types/index';
-import {TAGS} from '../utils/macromolecule/consts';
-import {SplitterFunc} from '../utils/macromolecule/types';
-import {getSplitter} from '../utils/macromolecule/utils';
+import {ISeqSplitted, SplitterFunc} from '../utils/macromolecule/types';
+import {UnitsHandler} from '../utils/units-handler';
+import {splitAlignedSequences} from '../utils/splitter';
 
 export function encodeMonomers(col: DG.Column): DG.Column | null {
   let encodeSymbol = MONOMER_ENCODE_MIN;
   const monomerSymbolDict: { [key: string]: number } = {};
-  const units = col.tags[DG.TAGS.UNITS];
-  const sep = col.getTag(TAGS.separator);
-  const splitterFunc: SplitterFunc = getSplitter(units, sep);
+  const uh = UnitsHandler.getOrCreate(col);
+  const splitterFunc: SplitterFunc = uh.getSplitter();
   const encodedStringArray = [];
   for (let i = 0; i < col.length; ++i) {
     let encodedMonomerStr = '';
     const monomers = splitterFunc(col.get(i));
-    monomers.forEach((m: string) => {
+    for (const m of monomers) {
       if (!monomerSymbolDict[m]) {
         if (encodeSymbol > MONOMER_ENCODE_MAX) {
           grok.shell.error(`Not enough symbols to encode monomers`);
@@ -31,21 +30,20 @@ export function encodeMonomers(col: DG.Column): DG.Column | null {
         encodeSymbol++;
       }
       encodedMonomerStr += String.fromCodePoint(monomerSymbolDict[m]);
-    });
+    }
     encodedStringArray.push(encodedMonomerStr);
   }
   return DG.Column.fromStrings('encodedMolecules', encodedStringArray);
 }
 
 export function getMolfilesFromSeq(col: DG.Column, monomersLibObject: any[]): any[][] | null {
-  const units = col.tags[DG.TAGS.UNITS];
-  const sep = col.getTag('separator');
-  const splitterFunc: SplitterFunc = getSplitter(units, sep);
+  const uh = UnitsHandler.getOrCreate(col);
+  const splitter: SplitterFunc = uh.getSplitter();
   const monomersDict = createMomomersMolDict(monomersLibObject);
   const molFiles = [];
   for (let i = 0; i < col.length; ++i) {
     const macroMolecule = col.get(i);
-    const monomers = splitterFunc(macroMolecule);
+    const monomers = splitter(macroMolecule);
     const molFilesForSeq = [];
     for (let j = 0; j < monomers.length; ++j) {
       if (monomers[j]) {
@@ -63,9 +61,8 @@ export function getMolfilesFromSeq(col: DG.Column, monomersLibObject: any[]): an
 }
 
 export function getMolfilesFromSingleSeq(cell: DG.Cell, monomersLibObject: any[]): any[][] | null {
-  const units = cell.column.tags[DG.TAGS.UNITS];
-  const sep = cell.column!.getTag('separator');
-  const splitterFunc: SplitterFunc = getSplitter(units, sep);
+  const uh = UnitsHandler.getOrCreate(cell.column);
+  const splitterFunc: SplitterFunc = uh.getSplitter();
   const monomersDict = createMomomersMolDict(monomersLibObject);
   const molFiles = [];
   const macroMolecule = cell.value;
@@ -150,4 +147,69 @@ export async function getMonomerLibHelper(): Promise<IMonomerLibHelper> {
 
   const res: IMonomerLibHelper = (await funcList[0].prepare().call()).getOutputParamValue() as IMonomerLibHelper;
   return res;
+}
+
+/** Calculates chemical similarity between reference sequence and list of sequences.
+ * Similarity is computed as a sum of monomer similarities on corresponding positions. Monomer similarity is calculated
+ * based on Morgan fingerprints.
+ * @param {DG.Column<string>[]} positionColumns List of position columns containing monomers.
+ * @param {string[]} referenceSequence Reference sequence.
+ * @returns {Promise<DG.Column<number>>} Column with similarity values. */
+export async function sequenceChemSimilarity(sequenceCol: DG.Column<string>, referenceSequence: ISeqSplitted): Promise<DG.Column<number>>;
+export async function sequenceChemSimilarity(positionColumns: DG.Column<string>[], referenceSequence: ISeqSplitted): Promise<DG.Column<number>>;
+export async function sequenceChemSimilarity(positionColumns: DG.Column<string>[] | DG.Column<string>,
+  referenceSequence: ISeqSplitted): Promise<DG.Column<number>> {
+  if (positionColumns instanceof DG.Column)
+    positionColumns = splitAlignedSequences(positionColumns).columns.toList();
+
+  const libHelper = await getMonomerLibHelper();
+  const monomerLib = libHelper.getBioLib();
+  // const smilesCols: DG.Column<string>[] = new Array(monomerCols.length);
+  const rawCols: {categories: string[], data: Uint32Array, emptyIndex: number}[] = new Array(positionColumns.length);
+  const rowCount = positionColumns[0].length;
+  const totalSimilarity = new Float32Array(rowCount);
+
+  // Calculate base similarity
+  for (let position = 0; position < positionColumns.length; ++position) {
+    const referenceMonomer = referenceSequence[position];
+    const referenceMol = monomerLib.getMonomer('PEPTIDE', referenceMonomer)?.smiles ?? '';
+
+    const monomerCol = positionColumns[position];
+    const monomerColData = monomerCol.getRawData() as Uint32Array;
+    const monomerColCategories = monomerCol.categories;
+    const emptyCategoryIdx = monomerColCategories.indexOf('');
+    rawCols[position] = {categories: monomerColCategories, data: monomerColData, emptyIndex: emptyCategoryIdx};
+    if (typeof referenceMonomer === 'undefined')
+      continue;
+    
+    // Calculating similarity for 
+    const molCol = DG.Column.fromStrings('smiles',
+      monomerColCategories.map((cat) => monomerLib.getMonomer('PEPTIDE', cat)?.smiles ?? ''));
+    const _df = DG.DataFrame.fromColumns([molCol]); // getSimilarities expects that column is in dataframe
+    const similarityCol = (await grok.chem.getSimilarities(molCol, referenceMol))!;
+    const similarityColData = similarityCol.getRawData();
+
+    for (let rowIdx = 0; rowIdx < rowCount; ++rowIdx) {
+      const monomerCategoryIdx = monomerColData[rowIdx];
+      totalSimilarity[rowIdx] += referenceMonomer !== '' && monomerCategoryIdx !== emptyCategoryIdx ?
+        similarityColData[monomerCategoryIdx] :
+        referenceMonomer === '' && monomerCategoryIdx === emptyCategoryIdx ? 1 : 0;
+    }
+  }
+
+  for (let similarityIndex = 0; similarityIndex < totalSimilarity.length; ++similarityIndex) {
+    let updatedSimilarity = totalSimilarity[similarityIndex] / referenceSequence.length;
+    for (let position = 0; position < positionColumns.length; ++position) {
+      const currentRawCol = rawCols[position];
+      if ((position >= referenceSequence.length && currentRawCol.data[similarityIndex] !== currentRawCol.emptyIndex) ||
+        (currentRawCol.data[similarityIndex] === currentRawCol.emptyIndex && position < referenceSequence.length)) {
+        updatedSimilarity = DG.FLOAT_NULL;
+        break;
+      }
+    }
+    totalSimilarity[similarityIndex] = updatedSimilarity;
+  }
+
+  const similarityCol = DG.Column.fromFloat32Array('Similarity', totalSimilarity);
+  return similarityCol;
 }