+++ packages/ApiTests/scripts/python/single_df_test.py
@@ -1,0 +1,7 @@
+#name: Python Single Df
+#description: df input/output
+#language: python
+#input: dataframe df
+#output: dataframe result
+
+result = df