+++ packages/UsageAnalysis/src/usage-widget.ts
@@ -2,7 +2,7 @@
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 import * as grok from 'datagrok-api/grok';
-import {usageAnalysisApp} from './package';
+// import {usageAnalysisApp} from './package';
 
 export class UsageWidget extends DG.Widget {
   caption: string;