+++ libraries/compute-utils/CHANGELOG.md
@@ -1,3 +1,14 @@
 # compute-utils changelog
 
-## 1.12.3 (WIP)
+## 1.13.5 (2023-08-10)
+
+### Features
+
+- Heavily optimized export functions for RFV
+- RFV now supports format setting for float inputs using `format` tag
+- RFV now supports precision setting for float outputs using `precision` tag`
+
+### Fixed bugs
+
+- RFV now saves selected output tab on function rerun
+- RFV now fully hides output during function run panel if no input tabs is provided